+++ src/TBD_macros.jl
@@ -672,12 +672,54 @@ function final_collect(sqlquery::SQLQuery, ::Type{<:snowflake})
     result = execute_snowflake(sqlquery.db, final_query)
     return DataFrame(result)
 end
+#using TidierDB
+function stream_collect(sqlquery::SQLQuery)
+    final_query = finalize_query(sqlquery)
+    res = DBInterface.execute(sqlquery.db, final_query, DuckDB.StreamResult)
 
-macro collect(sqlquery)
+    # Helper function to get the non-Missing type from a Union{Missing, T}
+    function non_missing_type(T)
+        T === Missing && return Any
+        T <: Union{Missing} ? non_missing_type(Base.typesplit(T)[2]) : T
+    end
+
+    # Initialize DataFrame with correct types
+    df = DataFrame([name => Vector{non_missing_type(t)}() for (name, t) in zip(res.names, res.types)])
+
+    while true
+        chunk = DuckDB.nextDataChunk(res)
+        chunk === missing && break  # All chunks processed
+
+        for (col_idx, col_name) in enumerate(res.names)
+            # Convert DuckDB data to Julia data
+            duckdb_logical_type = DuckDB.LogicalType(DuckDB.duckdb_column_logical_type(res.handle, col_idx))
+            duckdb_conversion_state = DuckDB.ColumnConversionData([chunk], col_idx, duckdb_logical_type, nothing)
+            col_data = DuckDB.convert_column(duckdb_conversion_state)
+            
+            # Append the data to the DataFrame
+            append!(df[!, col_name], col_data)
+        end
+
+        DuckDB.destroy_data_chunk(chunk)
+    end
+
+    return df
+end
+
+
+"""
+$docstring_collect
+"""
+macro collect(sqlquery, stream = false)
     return quote
         backend = current_sql_mode[]
         if backend == duckdb()
-            final_collect($(esc(sqlquery)), duckdb)
+            if $stream
+                println("streaming")
+                stream_collect($(esc(sqlquery)))
+            else
+                final_collect($(esc(sqlquery)), duckdb)
+            end
         elseif backend == clickhouse()
             final_collect($(esc(sqlquery)), clickhouse)
         elseif backend == sqlite()