+++ src/docstrings.jl
@@ -1051,7 +1051,8 @@ const docstring_db_table =
 
 `db_table` starts the underlying SQL query struct, adding the metadata and table. 
 
-#arguments
+# Arguments
+
 `database`: The Database or connection object
 `table_name`: tablename as a string. Table name can be a name of a table on the database or paths to the following types
       -CSV  
@@ -1060,8 +1061,9 @@ const docstring_db_table =
       -Iceberg
       -Delta
       -S3 tables from AWS or Google Cloud 
-`delta`: must be true to read delta 
-`iceberg`: must be true to read iceberg
+      - vector of CSV or Parquet paths to read multiple at once
+`delta`: must be true to read delta files
+`iceberg`: must be true to read iceberg finalize_ctes
 
 # Example
 ```julia
@@ -1086,3 +1088,41 @@ TidierDB.SQLQuery("", "df_mem", "", "", "", "", "", "", false, false, 4×4 DataF
    4 │ percent  DOUBLE               1  df_mem, false, DuckDB.Connection(":memory:"), TidierDB.CTE[], 0, nothing)
 ```
 """
+
+const docstring_collect =
+"""
+    @collect(sql_query, stream = false)
+
+`db_table` starts the underlying SQL query struct, adding the metadata and table. 
+
+# Arguments
+- `sql_query`: The SQL query to operate on.
+- `stream`: optional streaming for query/execution of results when using duck db. Defaults to false
+# Example
+```julia
+julia> db = connect(duckdb());
+
+julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9], 
+                        groups = [i % 2 == 0 ? "aa" : "bb" for i in 1:10], 
+                        value = repeat(1:5, 2), 
+                        percent = 0.1:0.1:1.0);
+                        
+julia> copy_to(db, df, "df_mem");
+
+julia> @collect db_table(db, "df_mem")
+10×4 DataFrame
+ Row │ id       groups   value   percent  
+     │ String?  String?  Int64?  Float64? 
+─────┼────────────────────────────────────
+   1 │ AA       bb            1       0.1
+   2 │ AB       aa            2       0.2
+   3 │ AC       bb            3       0.3
+   4 │ AD       aa            4       0.4
+   5 │ AE       bb            5       0.5
+   6 │ AF       aa            1       0.6
+   7 │ AG       bb            2       0.7
+   8 │ AH       aa            3       0.8
+   9 │ AI       bb            4       0.9
+  10 │ AJ       aa            5       1.0
+```
+"""