+++ NEWS.md
@@ -1,5 +1,9 @@
 # TidierDB.jl updates
 
+## v0.3. - 2024-07-28
+- adds support for reading from multiple files at once as a vector of paths in `db_table` when using DuckDB
+- adds streaming support when using DuckDB when `@collect(stream = true)`
+
 ## v0.3. - 2024-07-25
 - Introduces package extensions for:
     - Postgres, ClickHouse, MySQL, MsSQL, SQLite, Oracle, Athena, and Google BigQuery